[Reporting/Test] Split API test code under "usage"
